refactor: group session implementation into a session/ subpackage
